It's time to change
This commit “unpublishes” a number of my essays that touch on my
feelings about inequality and diversity in tech.

This, of course not a way of erasing history. The text of these posts
is still publicly available in this repository. Mirrors of the posts
are still in the Wayback Machine. You can peruse this commit with
GitHub’s rendered prose diff.

But I have decided not to openly host them any more. It’s time to move
on and try to use what little time I have left in different ways. I am
not changing my mind about where I want to go, but perhaps it’s time to
travel a different road to get there.

People sometimes call such behaviour “rage-quitting.” I don’t think I’m
angry, I’m quite pleased about a very large number of things that my
writing has done for me and for those who read it.

But every company retires some of its products as it evolves. Artists
change styles. Some sell their back catalogue, some burn their old
paintings as a form of catharsis, to stimulate new thinking and new
creativity.

That’s what I’m trying to do, to stimulate new thinking and new
creativity. So… Out with the old today, and hopefully in with the new
tomorrow.